requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<4.1.0
python-multipart>=0.0.6
python-docx>=1.1.0
aiosqlite>=0.19.0
jupyter-client>=8.6.0
ipykernel>=6.29.0
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
openpyxl>=3.1.0
matplotlib>=3.7.0,<3.10.0
seaborn>=0.13.0,<0.14.0
scipy>=1.10.0,<1.14.0
scikit-learn>=1.3.0,<1.6.0
imbalanced-learn>=0.12.0,<0.13.0
joblib>=1.3.0
xgboost>=2.0.0,<2.1.0
onnxruntime>=1.16.0,<1.20.0
pillow>=10.0.0,<12.0.0
opencv-python-headless>=4.8.0,<4.11.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
